Move state from ProductInventory component up to MainScectionContent component to better manage the application state, pass down data from MainSectionComponent to ProductInventory component via props, moved render of AddNewProductComponent from ProductInventory.js to MainSectionComponent.js
