+++ src/components/AddNewProduct/AddNewProduct.css
@@ -3,6 +3,8 @@
   padding: 0.938rem 2rem 0 0;
   justify-content: space-between;
   align-items: center;
+  margin-left: 2.188rem;
+  margin-right: 2rem;
 }
 
 .inventory-heading {
