+++ src/components/AddNewProduct/AddNewProduct.js
@@ -1,12 +1,12 @@
 import './AddNewProduct.css'
 
-// Modal initialization here onClick of New Product button will open Modal form.
-
-const AddNewProduct = () => {
+const AddNewProduct = ({ handleOpenModal }) => {
   return (
     <section className='add-new-product-container'>
       <h2 className='inventory-heading'>Inventory</h2>
-      <button className='new-product-button'>New product</button>
+      <button className='new-product-button' onClick={handleOpenModal}>
+        New product
+      </button>
     </section>
   )
 }
