+++ src/components/NewProductModal/NewProductModal.js
@@ -1,0 +1,8 @@
+const NewProductModal = () => {
+  // create a controlled component to handle the form data.
+  //on save of the form I need to pass the data back up to the MainSectionContent parent and add the product object to the list of existing products and render the new product listed
+
+  return <div>NewProductModal Component</div>
+}
+
+export default NewProductModal
